Cap18

Documentation for Cap18.

Cap18.DLResult

Cap18.Envelopes

Cap18.Reactions

Cap18.StiffnessLoadData

Cap18.Table4A

Cap18.change_string_by_slicing

Cap18.get_input_line

Cap18.parse_to_string

Cap18.DLResult — Type

struct DLResult

A result from the Table4A table (just one entry).

Fields

sta::Int - station along bent

dist::Float64 - actual distance along bent

defl::Float64 - deflection of bent at station

moment::Float64 - moment of bent at station

shear::Float64 - sheat of bent at station

Cap18.Envelopes — Type

struct Envelopes

A result from Table6 table (just one entry).

Fields

sta::Int - station along bent

dist::Float64 - actual distance along bent

max_moment::Float64 - envelope of max positive moment

min_moment::Float64 - envelope of max negative moment

max_shear::Float64 - envelope of max positive shear

min_shear::Float64 - envelope of max negative shear

Cap18.Reactions — Type

struct Reactions

A result from Table7 table (just one entry).

Fields

sta::Int - station along bent

dist::Float64 - actual distance along bent

max_reaction::Float64 - maximum reaction at support

min_reaction::Float64 - minimum reaction at support

Cap18.StiffnessLoadData — Type

struct StiffnessLoadData

A result from Table4 table (just one entry). Loads input as Cap and Stringer loads are applied directly to the cap. Loads input as Slab and Sidewalk Loads and Moving Loads are transmitted to the cap through the stringers, if any. Concentrated overlay loads are applied directly to the cap while distributed overlay loads are automatically transmitted to the cap through the stringers.

Fields

sta_from::Int - starting station

sta_to::Int - ending station

contd::Int -

cap_bending_stiffness::Float64 - bending stiffness of the cap

sidewalk_slab_loads::Float64 - sidewalk load (transmitted through stringers)

stringer_cap_loads::Float64 - stringer loads

overlay_loads::Float64 - overlay loads

movable_position_slab_loads::Float64 - movable loads

Cap18.Table4A — Type

struct Table4A

Table4A of the Cap18 analysis results.

Fields

results::Vector{DLResult} - collection of the dead load results from Table4A

Cap18.change_string_by_slicing — Method

change_string_by_slicing(original_string, start_index, end_index, new_substring)

A function to change parts of a string by slicing. This returns a new string. It is strict. The start and ending index range of the string will be replaced with the same number of characters. The function will error if more characters are given than the range provided. It will fill the left side of the range with spaces if the substring is smaller than the range provided.

Parameters

original_string - original string to modify start_index - index where modification starts end_index - index where modification ends new_substring - substring which will be inserted into string

Cap18.get_input_line — Method

get_input_line(
x::T, 
input_info::Dict{Symbol, Tuple{Int64, Int64}}, 
n::Int) where T

get_input_line(
x::T, 
input_info::Dict{Symbol, Tuple{Int64, Int64, Int64}}, 
n::Int) where T

This function takes a cap18 input struct, and outputs the line string for the .dat file that the CAP18.exe file uses.

Cap18.parse_to_string — Method

parse_to_string(x) = string(x)
parse_to_string(x::Bool) = string(convert(Int, x))

Parses an object type to a string type. For a boolean type, it parses as 1 or 0.